Move NetMessage (NetMessageBase) to net_message.h/cpp
