+++ src/Makefile.am
@@ -151,6 +151,7 @@ BITCOIN_CORE_H = \
   merkleblock.h \
   miner.h \
   net.h \
+  net_message.h \
   net_processing.h \
   netaddress.h \
   netbase.h \
@@ -271,6 +272,7 @@ libbitcoin_server_a_SOURCES = \
   dbwrapper.cpp \
   miner.cpp \
   net.cpp \
+  net_message.cpp \
   net_processing.cpp \
   node/coin.cpp \
   node/psbt.cpp \
